+++ packages/styled-components/src/sheet/Rehydration.js
@@ -1,12 +1,11 @@
 // @flow
 
-import { SC_ATTR, SC_ATTR_ACTIVE, SC_ATTR_VERSION, SC_VERSION } from '../constants';
+import { SPLITTER, SC_ATTR, SC_ATTR_ACTIVE, SC_ATTR_VERSION, SC_VERSION } from '../constants';
 import { getIdForGroup, setGroupForId } from './GroupIDAllocator';
 import type { Sheet } from './types';
 
 const SELECTOR = `style[${SC_ATTR}][${SC_ATTR_VERSION}="${SC_VERSION}"]`;
-const RULE_RE = /(?:\s*)?(.*?){((?:{[^}]*}|(?!{).*?)*)}/g;
-const MARKER_RE = new RegExp(`^${SC_ATTR}\\.g(\\d+)\\[id="([\\w\\d-]+)"\\]`);
+const MARKER_RE = new RegExp(`^${SC_ATTR}\\.g(\\d+)\\[id="([\\w\\d-]+)"\\].*?"([^"]*)`);
 
 export const outputSheet = (sheet: Sheet) => {
   const tag = sheet.getTag();
@@ -34,7 +33,7 @@ export const outputSheet = (sheet: Sheet) => {
 
     // NOTE: It's easier to collect rules and have the marker
     // after the actual rules to simplify the rehydration
-    css += `${rules}${selector}{content:"${content}"}\n`;
+    css += `${rules}${selector}{content:"${content}"}${SPLITTER}`;
   }
 
   return css;
@@ -53,14 +52,14 @@ const rehydrateNamesFromContent = (sheet: Sheet, id: string, content: string) =>
 };
 
 const rehydrateSheetFromTag = (sheet: Sheet, style: HTMLStyleElement) => {
-  const rawHTML = style.innerHTML;
+  const parts = style.innerHTML.split(SPLITTER);
   const rules: string[] = [];
-  let parts;
 
-  // parts = [match, selector, content]
-  // eslint-disable-next-line no-cond-assign
-  while ((parts = RULE_RE.exec(rawHTML))) {
-    const marker = parts[1].match(MARKER_RE);
+  for (let i = 0, l = parts.length; i < l; i++) {
+    const part = parts[i].trim();
+    if (!part) continue;
+
+    const marker = part.match(MARKER_RE);
 
     if (marker) {
       const group = parseInt(marker[1], 10) | 0;
@@ -71,13 +70,13 @@ const rehydrateSheetFromTag = (sheet: Sheet, style: HTMLStyleElement) => {
         setGroupForId(id, group);
         // Rehydrate names and rules
         // looks like: data-styled.g11[id="idA"]{content:"nameA,"}
-        rehydrateNamesFromContent(sheet, id, parts[2].split('"')[1]);
+        rehydrateNamesFromContent(sheet, id, marker[3]);
         sheet.getTag().insertRules(group, rules);
       }
 
       rules.length = 0;
     } else {
-      rules.push(parts[0].trim());
+      rules.push(part);
     }
   }
 };